[flang] Address issue#125, flang-compiler/f18#115, and improve error recovery on previous issues.
Original-commit: flang-compiler/f18@98e26d6245e93e17520fc64be1eff3b4b7db5d9f
Reviewed-on: https://github.com/flang-compiler/f18/pull/130
Tree-same-pre-rewrite: false

+++ flang/lib/evaluate/expression.cc
@@ -218,7 +218,7 @@ typename CharacterExpr<KIND>::LengthExpr CharacterExpr<KIND>::LEN() const {
         if constexpr (std::is_same_v<
                           const typename CharacterExpr<KIND>::Constant &,
                           decltype(x)>) {
-          return LengthExpr{x.size()};
+          return LengthExpr{static_cast<std::uint64_t>(x.size())};
         } else {
           return LengthExpr{LengthExpr::Add{x.x->LEN(), x.y->LEN()}};
         }
